portfolio/requirements.txt
Django>=5.2,<5.3
wagtail>=7.1,<7.2
django-colorfield>=0.11.0
Pillow>=10.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
